+++ backend/canisters/community/impl/src/lib.rs
@@ -47,6 +47,7 @@ use types::{BotSubscriptions, CommunityId};
 use user_canister::CommunityCanisterEvent;
 use utils::env::Environment;
 use utils::idempotency_checker::IdempotencyChecker;
+use utils::migrated_user_ids::MigratedUserIds;
 use utils::regular_jobs::RegularJobs;
 
 mod activity_notifications;
@@ -648,6 +649,10 @@ struct Data {
     public_channel_list_updated: TimestampMillis,
     #[serde(default)]
     certified_transfers: CertifiedTransfers,
+    // The latest ids of migrated users, as looked up from the LocalUserIndex whenever a user's id is found to
+    // have changed
+    #[serde(default)]
+    migrated_user_ids: MigratedUserIds,
 }
 
 impl Data {
@@ -779,6 +784,7 @@ impl Data {
             moderation_flags: Timestamped::default(),
             idempotency_checker: IdempotencyChecker::default(),
             certified_transfers: CertifiedTransfers::default(),
+            migrated_user_ids: MigratedUserIds::default(),
             public_channel_list_updated: now,
         }
     }