+++ backend/canisters/user/impl/src/lib.rs
@@ -27,6 +27,7 @@ use user_canister::UserCanisterEvent;
 use user_core::{Community, GroupChat, User};
 use utils::env::Environment;
 use utils::idempotency_checker::IdempotencyChecker;
+use utils::migrated_user_ids::MigratedUserIds;
 use utils::regular_jobs::RegularJobs;
 
 mod crypto;
@@ -401,6 +402,10 @@ struct Data {
     // their users
     #[serde(default)]
     pub known_multi_user_canisters: HashSet<CanisterId>,
+    // The latest ids of migrated users, as looked up from the LocalUserIndex whenever a user's id is found to
+    // have changed
+    #[serde(default)]
+    pub migrated_user_ids: MigratedUserIds,
 }
 
 impl Data {
@@ -455,6 +460,7 @@ impl Data {
             local_user_index_event_sync_queue: BatchedTimerJobQueue::new(local_user_index_canister_id, true),
             idempotency_checker: IdempotencyChecker::default(),
             known_multi_user_canisters: HashSet::new(),
+            migrated_user_ids: MigratedUserIds::default(),
         }
     }